Move reusable parts of PullRequestInstance into forge
