+++ bots/pr/src/main/java/org/openjdk/skara/bots/pr/CheckWorkItem.java
@@ -140,7 +140,7 @@ class CheckWorkItem extends PullRequestWorkItem {
         var labels = new HashSet<>(pr.labels());
 
         // Filter out the active reviews
-        var activeReviews = PullRequestInstance.filterActiveReviews(allReviews);
+        var activeReviews = CheckablePullRequest.filterActiveReviews(allReviews);
         if (!currentCheckValid(census, comments, activeReviews, labels)) {
             if (labels.contains("integrated")) {
                 log.info("Skipping check of integrated PR");
@@ -151,7 +151,8 @@ class CheckWorkItem extends PullRequestWorkItem {
                 var seedPath = bot.seedStorage().orElse(scratchPath.resolve("seeds"));
                 var prInstance = new PullRequestInstance(scratchPath.resolve("pr").resolve("check"),
                                                          new HostedRepositoryPool(seedPath),
-                                                         pr,
+                                                         pr);
+                var checkablePr = new CheckablePullRequest(prInstance,
                                                          bot.ignoreStaleReviews());
                 CheckRun.execute(this, pr, prInstance, comments, allReviews, activeReviews, labels, census, bot.ignoreStaleReviews());
             } catch (IOException e) {
